fix(esp32_hal_adc): Set MAX_MEAS_NUM to 255
It looks like 2 samples are lost after after getting MAX_MEAS_NUM+1.
With the value 10, 2 samples are lost every 11 samples, which makes the
actual sampling frequency divided by 1.22 (11/9).

With MAX_MEAS_NUM set to 255 (the maximum), the actual sampling
frequency is closer to the requested one. It is only divided by 1.00787
(256/254).

With the adc_continuous example, the requested sampling frequency is 20
kH, but without this patch, the actual frequency is 16363 Hz (20000/(11/9)).

With this patch, the actual frequency is now 19843 Hz (20000/(256/254)).
Note that that frequency is achieved by:
 - Setting EXAMPLE_READ_LEN to 1024
 - Setting .max_store_buf_size to 4096
 - Not printing all ADC values on the serial

Also note that this explains why the example in v4.4 (MAX_MEAS_NUM
configured to 250) seems to work way better than in v5.2 (MAX_MEAS_NUM
fixed to 10)

Signed-off-by: Detlev Casanova <dc@detlev.ca>